Properly handle all SendEvent results
I added config for all the statuses of the Swift_Events_SendEvent (including spooled https://github.com/swiftmailer/swiftmailer/pull/556)
This fixes an issue, where all non-SUCCESS result codes are logged with ERROR level (assumed NON_SUCCESS).
Else log fills up with ERRORs for every message that is spooled first before real sending. If we lower the level then we can not distinguish the real errors.
Also TENTATIVE status does not always mean that the sending failed. So log level for it should be configurable as well.

also increased the visibility of private properties and log method so it is easier to extend (or decorate) the class